Initial mips implementation working

+++ src/commonMain/kotlin/com/tobi/mc/mips/MipsArgument.kt
@@ -5,5 +5,5 @@ sealed class MipsArgument {
     data class Register(val name: String) : MipsArgument()
     data class IndirectRegister(val name: String, val offset: Int): MipsArgument()
     data class Label(val label: String): MipsArgument()
-    data class Value(val value: Long): MipsArgument()
+    data class Value(val value: Int): MipsArgument()
 }
